Feature/skill rework (#4447)
New Skill System:

Instead of the 5 basic skills there are now 11 skills.
10 normal skills that are used for the jobs

Intelligence which is the main driver of the leveling speed
The only way to level intelligence is at the library or when the citizens are still children at a school (next building to be implemented)

Skills are now leveled and skills themselves influence the efficiency of the workers.

Every job requires 2 of them. Those two make him work "better" and he levels up those two skills while he goes

Every skill has a complementary (Gets another 1% of xp) and a contrary skill (loses 1%)

This PR simplifies the leveling system as well by moving the skills in the building and removing the load from the view and AIs to calculate the modifiers.