Stash is added
This commit contains the logic behind the worker hut, the block and the TileEntity.
I've created a TileEntity as I've found no pratical way to change the `inventory` from `AbstractTileEntityRack` for all instances of Stash. This obviously can be changed if a better way comes up. For now, it's a protected field and has no way to be re-assigned, so this is what I came up.