Feature/school (#4585)
This PR adds a school to the colony.
School give work to one teacher total and up to 2 kids per level.
When kids grow up the automatically leave the school.

Kids only stay in school until noon, then they roam around freely again.

In school kids sit around on carpet and "study". Randomly they get a recess and run around freely in the building.

To compensate kids that study in a school grow up twice as quickly in the remaining time to compensate.

Teachers give out paper to kids and randomly sit down with kids to teach them.