Stash #2121 (#4545)
First, thanks for this great mod! I loved playing with it, so I wanted to give back a little!

Implements #2121

Changes proposed in this pull request:
Added a Stash, that will have items inside that the DeliveryMan will pickup and move to the Warehouse.
So, this incomplete PR is an attempt to implement #2121 , although just the first comment, as the second one I felt it was beyond the scope of the Stash, as I thought this should be kept to a minimal functionality.

Core Functionality
The Core functionality of how I thought was this: when an item gets deposited, it bumps the priority up to 10, to encourage the DeliveryMan to pickup items ASAP on them. When the DeliveryMan picks up the items, it drops the priority to a minimum. So the bumping and the dropping are being handled by the TileEntity and the Inventory Handler.

Found issues
I noticed that the dropping of the priority should be done automatically, but it's only done when it's the last building on the list? I didn't quite understand why this didn't work:

minecolonies/src/main/java/com/minecolonies/coremod/entity/ai/citizen/deliveryman/EntityAIWorkDeliveryman.java

Line 217 in 1fdde5f

 building.alterPickUpPriority(-10); 
Temporary work
For now, I don't have a fixed GUI. Suggestions are welcome for functionality.
Also, I don't have a model. If you guys have a suggestion, one to spare or a software for me to take a look (I will try Blockbench).
Also, I don't know how you guys manage your language files, so, for now, I still have no lang strings for it.

Conclusion
I'll update after I have done more work, I just wanted to say the work is being done, and get feedback if you think the direction is right. For now, only the first commit actually needs to be looked at 135da01

Review and suggestions are welcome