Feature/stripped lumberjack (#4925)
Feature Suggestion
The new (awesome) Asian building style requires a fair bit of stripped wood.
Stripped variants of woods cannot be crafted, but must be made by right-clicking woods/logs with an axe. Which becomes old really quickly when the buildings require hundreds of them.

It makes sense for the lumberjack to be able to craft stripped versions of logs.

Draft implementation
Lumberjack turns vanilla wood/log to stripped versions if a request for this exists in the request pipeline.
The lumberjack hut's GUI shows the list of recipes (i.e. the stripped versions), but you can't remove them and you also can't teach new recipes.