Allow certain requests to have a minimum count (#4968)
This PR basically replaces the previous handling where specific workers were able to make requests that were partially fulfilled with a request specific handling. Thus worker can request something with a count and a min count, where the RS will try to fulfill the count and if this is not possible but the system has a bit less it will try to match the minCount.

I also removed some of the heavy streams with quicker for loops.