Add a new reposition approach (#5220)
Reposition rework:

Instead of reposition this works now like this:

Building has button to deconstruct. Calls builder to deconstruct. After this is finished. You get a "Pick Up" button. You get the hut with a colony id and the level in it.
On placement with the buildtool that gets applied.