Add the rallying banner (#5217)
The rallying banner allows players to summon their guards to them.

Features
Sneak-Rightclicking on a guard building hut adds it / removes it from the rallying list.
Right-Clicking in the air opens the GUI, which allows the player to remove Guard Buildings from the list without sneak-rightclicking them.
Sneak-Rightclicking in the air or on the ground summons or dismisses the guards to the player.
While running to the player, guards will receive a massive speed boost (Speed2 at minimum, but up to Speed5 when highly trained).
While rallied, guards will glow in the colony's team color, and are immune to damage from the rallying player. The same is not true the other way around!
Guards will return to regular duty when dismissed, or when the banner is not in the player inventory anymore.
When an active banner is put back into the player inventory (grabbed from a chest, or moved from slot to slot), chances are that the rallying will proceed.
There is a new permission that allows players to rally guards of other people's colonies.
Limitations
Guards will use a high amount of food while rallied. Currently it's calibrated to reduce the guard to 2 saturation points after 5.5 minutes of rallying.
Dismissing the guard while below 2 saturation will remove those 2 saturation, forcing the guard to eat.
Guards will only rally within the colony area. Attempting to move them out of the colony will dismiss them automatically.