Tavern2 (#5355)
Adds a new feature which spawns visitors to the tavern, which a player then can recruit for varying costs.
For that it adds a new interaction type called recruitmentinteraction, which displays and handles the item stack cost for recruitment. The costs for recruitment are loaded from the config, with given item and weight on it.
Interactions now are able to return the icon supposed to render for it.
Interaction responses now are kept in insertion order instead of random hashmap order
Interactions now know the player its interaction it, rather than the world.
Added a WorldUtil class with helper functions for checking chunk load status properly
Added a schematic constants class, for string tags used in schematics. Should be in sync with the builder's tag list.

Fixes:
Fix bed pos and last pos getting written to the same nbt tag, overwriting eachother.
Fix guards not moving out when fired.
Fix tavern music only playing when you were close enough when it started
Fix raider respawns, when sometimes an entity gets into an unloaded chunk but not removed
Fix claim command requiring an existing colony, so you now can unclaim already deleted colonies
Fix whereami command distance output
Fix loading chunks on entity enter event