Update README (#5575)
The instructions provided for compiling and setting up the mod are invalid.
With the update to ForgeGradle 3, the setupDecompWorkspace command was removed, in favor of Gradle's own initialisation system. Any attempts to use this command will error in newer versions.

This PR changes the README.md to reflect these changes.

Additionally, some tersing of long sentences was included, breaking them up with commas or moving text around to improve the flow.