Rework the baker to be a real smeltercrafter (#5785)
The baker is now a true crafter, and will bake on-demand
The baker no longer bakes as long as it has resources, but now supports prestock to force baking.
Baker recipes reworked, and now have doughs that the baker crafts, and then cooks in the ovens to produce final result.
Reworked the way all the smeltercrafters handle fuel management. It is much faster and more reliable now.
Fixed quite a few small issues across the crafting and display code, especially around NBT handling - This will better support potions in the future.