Break up big crafting jobs (#5852)
Break crafting jobs into chunks based on the ingredients fitting in the inventory of the crafter
This resolves the problem with cases where a crafting job requires more slots to load the recipe than the citizen has available. this becomes very visible when the recipe uses an ingredient with a maxStackSize of 1, and the output stacks higher than the number of inventory slots on the citizen.

Today, we have a special recipe for level 3 bakery that uses water bottles, and making a stack of that recipe needs 32 of those, with a total slot requirement of 33 + 1 for output.

The solution is to recognize the max recipe iterations we can load inputs for, and instead of creating a single crafting job of the full amount, calculate the maximum size we can process, and break it up into batches of that or smaller to accomplish the task.