Rewrite of door/gate/trapdoor toggle AI (#5886)
Rewrite of door/gate/trapdoor toggle AI, this lets us use the same AI for blocks with similar toggle behaviour.
The idea is the same as the fenceGateAI before: we check the parts of path for toggleable blocks in range, and save their positions(previously we could only use one position). If we found some blocks the AI becomes active, while active it randomly activates one of the blocks it found. With multiple blocks in range it eventually finds a combination which allows it to go through.

Pathing now allows single trapdoor blocks as passable.
Pathing now also adds the starting node to the path, previously only nodes with a parent were added.

Example of what they now make their way through: