Adding split networking. (#5885)
This adds splits networking.
Our packets are getting to big, even with string reductions on the RS side of things packets that we are sending around are getting extremely big for large colonies.

This PR adds a wrapping system which automatically splits all packets that are bigger then 90% of the max value.

On the receiving side the packets are processed, stitched back together and then processed as normal.