Delivery orders (#5921)
This adapts the way deliveries are assigned to deliveryman to a multi-factor comparison system.
The resolver for dmen now resides in a warehouse and resolves the request it gets to its delivermen in our case those of a single warehouse. So now we compare all assigned deliverymen of a warehouse to get the best fit for the request.
The resolver now ask the dmen for a score for adding a certain request and then compares the score to other dmen's scores. The score(lower is better) itself is combined of these factors:

Closeness: How close the target/source of the new request is compared to an existing one, in relation to its total travel length.
Priority(also contains age): Overtaking a request of a higher priority is worse.
Pickup/Delivery alternating: When alternating between pickups and deliveries with close targets the score gets a little bonus.
Index: The index we would add the new request at, higher the more request we overtake and the more request we have in total.
These result in lists on deliverymen of approx ~3 times closer deliveries, allowing the AI to also bunch up deliveries more nicer in the future.

Fixes visitors sometimes getting spawn on zero pos.
Fix raids not respecting distance to unloaded buildings
Fix shipsraids sometimes instantly despawning after killing spawners, now takes a day to do so.