Multi-Output recipes, and reliable secondary outputs (#5905)
Add support for multi-output recipes, so that one recipe can create all variants of one timber block
Add support for custom recipes to require that the recipe 'exists' to be valid
This enables a multi-output recipe to replace a classic recipe in which the inputs match
When a multi-output replaces, it also looks for classic recipes who have outputs that are alternate in the multi-recipe and removes them
Add support for multi-recipes to cycle through the outputs in the recipe list in buildings
Add ability to capture actual secondary outputs (stay in grid) from recipes as they are being taught, This increases compatibility with many mod recipes, including aquaculture fish fillets.
Some minor tweaks to display of the request tree in buildings, to make crafting jobs a bit clearer
Fix Combat Academy and Archery to mark beds as unoccupied on wake up, letting the residents sleep the next night
use Localizable strings for Combat Academy and Archery
Ensure that multi recipes replace simple recipes, and that they update when they change
Change all of the lumberjack recipes to be multi-recipes. They are now Log -> [Stripped Log, Wood, Stripped Wood]