Adds 5 new scroll items with unique effects (#6169)
Adds 5 new scroll items with unique effects:
A colony home teleportation scroll, which teleports the user to the registered colony.
A colony home area teleportation scroll which teleports the user and nearby players to the colony.
A regeneration scroll, granting regeneration to the player and citizens around him, found in chests.
A guard help scroll which temporarily summons the guards of a registered guard building to the user on follow mode.
A worker highlight scroll which can be used on a worker building an highlights the worker with a glow effect and gives him a slight speed effect too.