Low-hanging Fruit: UI improvements, off-hand build wand, crafter recipe adds, and expanding raid command capabilities. (#6288)
Adds a progress message at 25 citizens, pointing players to the University research.
Adds disabled/enabled button functionality and tooltips for Warehouse Upgrade and Sort.
Adds a time check to WAREHOUSE_FULL messages, to reduce message spamming, and gives different messages depending on whether the warehouse can still be upgraded (in which case it can't use emerald blocks) or not (in which case it can't be upgraded.
Adds glowstone dust as an input to the mechanic, and sets glowstone (block) as non-reducible to avoid potential surprises to datapack makers. Added a few other exclusions for storage-like blocks that Forge does not inherently tag as storage, for similar reasons.
Adds Blackstone and Polished Blackstone (and polished blackstone slab/bricks) and Polished Basalt as usable ingredients to the Stonemason.
Allows use of build tool in off-hand.
Removes duplicate sendPlayersMessage for WorkOrderBuildRemovals.
Adds a message on completion of a deconstruction work order to warn players to use the Pick Up botton.
Add a third count to the Town Hall's citizen summary page, indicating the maximum limit of colonists from research.
Adds an optional argument to raid-now, raid-now-all, raid-tonight, and raid-tonight-all to force the raid to be a specific type.
Adds a raid event for pirates without a ship, only summonable by manual raid commands for now.