Improves logic, comments, and clarity of WindowResearchTree.java.  Removes uses of a hard-coded 6, and reduces the usages of MAX_DEPTH when possible to use building.getMaxBuildingLevel instead.  There are still a handful of usages, but changing them involves either cracking open protection levels on UniversityBuilding or using a different hard-coded variable.
Reduces extraneous looping in drawTreeBackground, gets rid of Math.min calls in drawTree (and does some cleanup related to them).

Fixes some behavior related to the armordurabilitymultiplier and tooldurabilitymultiplier effects, as discussed with SomeAddons in the "getting-started" discord.  Clarifies translation text for them.  This is effectively a small nerf to Diamond Coated's line, and a significant buff to Diamond Skin's line.