Rework the Sifter (phase 1) to be based on Recipes/Loottables  (#6316)
Simplified the sifter and reused the crafting system to support it.
Sifter no longer needs you to choose what to sift, it will sift anything siftable in it's building inventory.,
Sifter now supports prestock requests, which is the new way to request siftables to process
Sifter still supports a maximum number of blocks sifted per day, be sure to set this!
Sifter is now faster, and capable of being MUCH faster at higher levels.
Sifter will rarely at higher primary skill levels, and only with iron/diamond meshes, produce more than one item from a single block
Added support for diminishing returns on skill level, sifter is the first to use this for it's speed/luck
Early levels get a bonus, ie: skill 10 is an effective skill of 19, but skill 90 is effectively 99.
Added a chance for the baker's level 3 bread dough recipe to produce empty bottles.
This is only the first round of changes, and is effectively a drop in replacement as far as overall functionality. I plan on adding real items for the Sifter meshes, with durability that is effected by sifting. Will likely also add luck bonuses based on skills, and more complex loot tables (rare sand from gravel anyone?)

Phase 2 will likely require the new research system also, as it allows for more complex conditions to trigger recipe availability.

To be clear, Phase 2 will be a separate pull request after this one is in. This one stands alone.