Fix for Pupil Display, Baker/Bakery on Chorus Bread (#6361)
Changing counting approach for School-related workers (pupils and teachers) to avoid possible double-counting. Adds comments to improve clarity.
Changes the building research requirement for "Know The End" from "bakery" to "baker", to match the getSchematicName for the baker hut ("bakery" is only used in translation files). This'll be obviated with the research data pack, but this fix solves the problem for the meantime, and won't interfere with that PR.