Bug Squashes, Eastern Name Order, and Gender Issues
Moves logic related to the PVP-mode management of the TownHall block to BlockHutTownHall. Cleans up the sendPlayerMessages to insert arguments into TranslationTextComponents properly. Removes the PlayerEvent.BreakSpeed, which only had logic related to the TownHall block but was getting called for everything except blockhuts.

Fixes double-counting of enchantment levels on fishing rods.

Changes gendering for initial settlers spawned by the TownHall to spawn men if the town has more women, and to spawn women if the town has more men.  Also changes the first colonist to have a random chance of being male or female.  Pattern will normally be female-male-male-female or male-female-male-female.

Removes duplicate placeConstructionTape in addNewBuilding, which was causing it to fire for  IRSComponentBlocks like the postbox and twice for normal buildings.

Changes some model files to handle the 1.15.2+ Chest texture file change.  Warehouse, Shepherd, Chicken Farmer.  This could also be done for Swineherd and Courier, but they kinda look better as-is, since they're closed and the only effect is a shaded spot on the top of their chests.  Did fix the under-latch shadow textures for the Courier.

Sets ItemTagResolvers to give the translated text for items that meet their requirements instead of the long Minecraft name of the tag (#com.minecraft.items.tags.Flowers).  Only impacts Beekeeper for now.

Adds a configuration option for Eastern Name Order.  For now, just switches to surname+lastname without space.  There are some Asian cultures that use more complicated structures and a whole mess of complications from social norms, but this'll handle the simple good-enough version.