Bug Squashes, Eastern Name Order, Gender Issues, and ItemTagResolver Display Tweaks (#6408)
Moves logic related to the PVP-mode management of the TownHall block to BlockHutTownHall. Cleans up the sendPlayerMessages to insert arguments into TranslationTextComponents properly. Removes the PlayerEvent.BreakSpeed eventhandler, which only had logic related to the TownHall block but was getting called for the vast majority of blocks but not BlockHuts.
Fixes double-counting of enchantment levels on fishing rods.
Changes gendering for initial settlers spawned by the TownHall to spawn men if the town has more women, and to spawn women if the town has more men. Also changes the first colonist to have a random chance of being male or female. Pattern will normally be female-male-male-female or male-female-male-female, but will recover from gender-seeking zombies a little more elegantly.
Removes duplicate placeConstructionTape in addNewBuilding, which was causing it to fire for IRSComponentBlocks like the postbox/stash (and twice for normal buildings).
Changes some model files to handle the 1.15.2+ Chest texture file change. Warehouse, Shepherd, Chicken Farmer. This could also be done for Swineherd and Courier, but they kinda look better as-is, since they're closed and the only effect is a shaded spot on the top of their chests. Did fix the under-latch shadow textures for the Courier.
Sets ItemTagResolvers to give the translated text for items that meet their requirements instead of the long Minecraft name of the tag (#com.minecraft.items.tags.Flowers). Only impacts Beekeeper for now. Not a perfect fix, as the StandardInteraction will show the first eligible item in the request (for vanilla flowers, dandelions), but solves the problem well enough without requiring deep changes.
Adds a configuration option for Eastern Name Order. For now, just switches to surname+lastname without space. There are some Asian cultures that use more complicated structures and a whole mess of complications from social norms, but this'll handle the simple good-enough version.