Port of Research-Datapack Fixes (#6430)
Fixes display of tooltips for hut blocks that require research to build.

MaxCitizens research will now go up to the maxCitizenPerColony limit, instead of 25 below, when at max level. This will also allow very low levels of maxCitizensPerColony (below 100) to apply at all. The higher ranks won't be able to adjust to match the server config settings yet, though, as remote clients don't get that information, but the workarounds to do so are very ugly and I don't want to mess with config file ranges for behavior that'll change again with the research data pack.

Changes TOOL_DURABILITY and ARMOR_DURABILITY modifiers to act as random > (1 / (1 + effectVal)), as discussed with someaddons in the #getting_started chat. This is effectively a small nerf to the Diamond-Coated research chain and a large buff to the Diamond Skin research chain (whose highest two tiers would only really show up if you tried to zerg rush a Wither with guards). May eventually be worth doing a more focused balance pass.

Allows teachers to apply their PrimarySkill (Knowledge) to XP gain of pupils regardless of "More Books" research status.

Fixes SATLIMIT research to reduce the saturation required for citizens to get double healing, instead of increasing it above the possible range.

Fixes costs of the Deep Pockets- Loaded - Heavily Loaded research, at least for new worlds or when a config file is rebuilt.

Cleans up a few effect spelling or capitalization oddities, where doing so is non-breaking.