More sifter improvements (#6529)
Add icons of possible drops defined in secondaryoutputs when primaryoutputs is empty in display of recipes
Add support for more context in loot tables, but don't enable it yet. (This allows loot tables to use mob loot tables)
The additional context is intended to be used in the nether resource gatherer, and possibly even a rudimentary expedition hut, but it's not used by normal crafting. I will likely add a property to crafters that can be used to enable it in the future, or it will be called directly similar to the way the sifter calls today.