Improve Research Branch Customizability (#6722)
Changes the WindowUniversity display to use a ScrollingList, so to better support 8+ research branches.
Adds support for the branch-type, subtitle, sortOrder, and hidden properties to branch-level jsons.
sortOrder and subtitle are similar to their effects in research jsons. Branches with a higher sortOrder will be placed lower on the University Window, branches with a subtitle will have a tooltip on the selection list.
hidden will prevent the display of a research branch if all of its primary researches are set to hidden, and none of those researches have their requirements fulfilled. Branches where all primary researches are set to hidden but the branch itself is not will display a disabled ButtonImage with the text marked out and a tooltip listing the unlock options.
branch-type allows wider styling or logical changes to a full branch. The only currently supported types are "default" (current behavior) and "unrestricted-tree", which is meant to support Mekle's use case where a branch needs University-specific capabilities not readily managed through vanilla advancements (either by having colony-wide effects or requirements, possibly occurring when players are offline, needing strictly enforced hierarchy, or wanting to require an item submission) by removing the time and university-level limitations. Down the road, I could see potential uses for a couple stylings that use nested ScrollingList instead of a tree-style dragzoomview, and there are likely other more esoteric uses.
Cleans up the related implementation logic to use Actual Objects instead of parallel hashmaps.
Fixed "Heavily Loaded" not properly having the immutable property