Research UI Updates (#6890)
Fixes play/cancel/check icons on the research UIs not accepting clicks.
Fixes the time labels for the tree gui's top showing incorrectly calculated hour estimates.
Extends the research name label, at least enough for Captain of the Guard and What is this Speed.
Changes text color for unfulfilled requirements and warning text to a lighter salmon pink instead of the dark red.
Overrides hovertext on research requirement item icons, to more precisely describe the requirement and change textcolor to salmon/aqua depending on whether it is fulfilled/unfulfilled.
Right now, this includes building and research requirements on the main tooltip for the research, but item requirements only show on mouseover for their specific icons. I'm not sure that's the right compromise between clarity and too-much-text; it's pretty easy to change either direction, but a quick attempt at including everything made multi-item researches look a little much.