New Graveyard Building - New Graves - New Undertaker Job (#6737)
Changes proposed in this pull request:
Add a new building: the graveyard
Add a new job: the gravedigger
When citizen die within the city border; spawn a grave that contain the citizen info & its inventory
This grave will decay after a certain amount of time an go to a decay state. After a while the decayed grave will disappear
The player can retrieve the inventory of the grave
The gravedigger role is to collect graves of dead citizen, bring their inventory back to the graveyard, bring the grave back to the graveyard and build a new gravestone (named grave) in the graveyard if there is space for it
The gravedigger has a very small chance of resurrecting a citizen before burying them. This change improve with the graveyard level, the mana skill of the gravedigger and 2 research in the university
This chance is capped to 10% with all maxed out bonus, but this cap can be slightly increase by building mystical site or with 2 research in the university
The gravestone built by the gravedigger in the graveyard have a custom renderer displaying the name and job of the buried citizen
The gravedigger is immune to mourning, so they can work when another citizen die
When a citizen is burry, the mourning time for this citizen will fall from 1 full day to half a day
add achievement for the first grave collected by the grave digger
add achievement for the first resurrection
the graveyard gui contains an information text that explain how it works
the graveyard gui contains a list of grave to collect in the city
the graveyard gui contains a list of all citizen burried in it
When the gravedigger has no grave to collect, he will visit home to get to know citizen and visit magical buildings in the city to train his mana skill
include all the stone schematics to allow players to already start using the feature. Our schematics builder can feel free to modify them of course.
The graveyard schematics should contain tile entity "Named Grave" to specify to the gravedigger where to build named grave