Medieval schematics Update - Alternative Dark Oak
This is it, finally medieval styles got properly updated and finished since the version jump into 1.12 that changed all.
Updated medieval schems, mainly alternative dark oak:
(These buildings got reworked and need to be deconstructed/rebuilt)
-Townhall
-archery
-baker
-combatacademy
-barracks and its towers
-restaurant
-deliveryman
-builder
-forester
-sawmill
-swineherder
-chickenherder
-guardtower
-fisherman
-warehouse
-cowhand
-farmer
-miner

--Added dark oak mystical site.

""Legends say that anyone thats builds the lvl5 darkoak medieval alternative townhall in survival deserves to be crown""

I have to thanks everyone that uses medieval schems, your comments and opinions as well as the fun you have with the styles is my main motivation to work on this. Thank you a lot as well as the awesome devs :)