Unlock item tooltip (#6943)
Adds a "show-tooltip" property to crafterrecipes.
Transfers crafterrecipes with either of these traits set to true to remote client, ~~and stores them inside a separate recipeOutputMap. Documents in code which recipe map is populated based on sidedness. ~~ This should help with some of the JEI plugin work,
Properly sets CustomRecipeManager to clear at start of data reload listener (eg, no longer having recipe data from previous worlds, or before a data pack disable).
Adds a client event handler to show tooltips for all items with the "show-tooltip" property, and to handle research unlocks for hut blocks.
Removes addInformation-override based tooltips for research on AbstractHutBlocks, and for crafter origin from baked goods and gates. Item-specific matters like effects and gate behavior remain in the item-specific addInformation call, but this should be more manageable for customized environments.
Cleans up some research-related code and comments.
Provides a way to collect the research (or researches) tied to a given effect. This may be null, if no research with the effect is loaded.
Provides a fast way to check if an item, ItemStack, or ItemStorage can be made with any crafterrecipe.
I'm not perfectly happy with the UI result, so feedback on that would be appreciated. Current tooltip behavior is:

Red text when in a colony that does not have a building of the required level, or the research, or when outside a colony when one's owned colony does not have the requirements.
Aqua text when within a colony that does have the research or required building of the required level, or when outside of a colony but when your owned colony has the research or required building of the required level. I'm not sure if it makes sense to have the research ones disappear when they're fulfilled, instead.
Gray text for simple information, such as where the recipe doesn't require research or a specific building level.
Similarly, not sure if getting an icolony based in proximity is the right choice, or if it should go by ownered colony, or only when within the colony borders. Updated re Raycoms.