ItemStorage fixes, Tooltips, Beekeeper_2 Research (#7006)
Fixes serialization issue in ItemStorageFactory causing Research Costs to not display properly.
Adds tooltips to the Town Hall and Colonist Happiness windows, describing the various Happiness modifiers.
Adds a maximum amount of stored additionalChildTime, currently set to the amount that will allow one child to be instantly-grown. Handles a case where the colony would store up enough additionalChildTime that every child past that point would be instantly-grown, only really occurring on servers where a player was leaving themselves logged in for long periods of time away from the colony.
Adds a research allowing Beekeepers to collect both combs and bottles, which alternates between the two. Various changes to the AI and message structure to support that. Sorry for stealing the assigned topic, Raven, but owed you given how long the Research system update took me. I'm not attached to the name or subtitle, but a Winnie the Pooh reference or two seems on-point.