Remove tool smelting (#7178)
This removes the feature of the smelter allowing them to "dismantle" tools, weapons and armor and receive raw materials back.
This was a feature that was very old and was an attempt in 1.12 to get people to use our smelter for ores as, most people only used tinkers. However, with the progress of our RS and the crafting improvements, this system could easily lead to infinite resource loops (Make a discounted iron chest plate at the blacksmith, and then dismantle it at the smeltery).
In light of the upcoming JEI support this would shine more light on this feature (almost no one uses and knows about) and as such I'd like to remove it before that gets merged.