Fix JEI datapack loot, kinda (#7224)
If we can't load a crafterrecipe's loot table (probably because it was in a datapack instead of a mod, and the loader can't cope with that just yet), fall back to showing the alternate outputs (which are set manually for use in the recipe list). This should show the correct items (if set properly in the recipe) but not the drop chances.
Also fix that missing else that briefly spawns a duplicate citizen.