Schematic updates from John (#7733)
Rescan of acacia, jungle and spacewars
Barrack towers have been converted to just be the hut block, the towers themselves are incorporated in the barracks now (like in lost cities)
This has the renovated jungle style, meaning most huts have to be placed in trees going forwards
Jungle has a new farm, which is placed in junglealternative. It would break existing jungle colonies otherwise.