Feature: fishing bonus loot tables (#7744)
Refactors the existing loot table datagen to use vanilla datagen (+ LootTableAnalyzer where needed) instead of LDT datagenerators library.
Now the only thing remaining which uses that library is the sounds datagen. I have not investigated how easy it would be to replace that too and remove the dependency entirely.
A bunch of datagen files have "changed" as a result of this, but it's just reformatting; unless I've overlooked something there's no actual behavioural change.
Moves the fisherman bonus loot generation from config file to datapack loot table. This makes it easier to tweak for modpack authors, among other benefits.
Intended change: instead of spontaneously appearing in the fisherman's inventory the loot now spawns in the world at the fishing bobber and is pulled towards the worker, exactly as the fish does.
Not-completely-intended change: while the base chance (at zero worker skill) for the extra drops is the same as before, the overall chance of bonus loot is slightly lower (e.g. at skill = 100, chance of a prismarine shard is 9.6% instead of 12.5%), due to the way that loot table scaling works. Not sure if this is still ok or if we want to tweak it further.
Not changed at all, but noticed in passing and it seems weird: since the fisherman is using a custom fishing bobber (presumbly since the vanilla one is hard-coded for players), it automatically fails the loot check for FISHING_TREASURE and this table is never used. (Consequently, enchanting a fishing rod with Luck of the Sea is a net loss (since it decreases the drop weights of non-treasure, but then doesn't add anything else to compensate; this also happens for player-used enchanted rods if not fishing "in open water", whatever that means).)
Also not changed, and not sure if intended: the config comment specified that sponges would not drop before Fisherman L4, but in actual fact everything starts dropping at L3. But there's separate loot tables now per hut level so it's easy to customise if we want to (or a modpack author wants to). The sponge will no longer drop from L3.
Also not changed, but seems probably intended: the fish/junk loot table is affected solely by the Luck of the Sea enchantment, while the bonus loot chance is affected solely by the worker's primary skill level.
Switches the standard fish loot to use custom tables (but these still delegate to the vanilla tables for the most part, so should get similar loot automatically).
This is primarily to allow the FISHING_TREASURE loot to drop, which was otherwise blocked by use of a custom bobber.
The condition for treasure loot to drop is different from vanilla -- here we simply check whether the bobber itself is in a biome classed as ocean, rather than the AABB block check performed for vanilla fishing. Note that this means it is very unlikely that fishing in a hut-integrated fishing lake will ever produce treasure loot, but that's better than never. (This also means that using Luck of the Sea is still counterproductive unless they are actually fishing in an ocean biome.)
Added JEI integration for the fisherman: