Fix following guards idling on player death
The problem was that the PlayerEntity object of the player to follow
before and after dying was not the same and it wasn't null. It seems
like the old PlayerEntity is sent to the player's spawn while the
player is given a new PlayerEntity to control, thus making the
building guards refer to the wrong entity.

This fix works by keeping track of the player's UUID (which doesn't
change after death) and resolving that into the correct PlayerEntity
object when needed.